PCB Files
I added all the PCB files needed to make a PCB for this project. I will probably be selling some Through Hole kits and some completed surface mount boards in the near future. But for now if you would like to make your own here is a pretty decent PCB setup for you. It was made with Kicad.